--- configurations/test_config.yaml
@@ -1,10 +1,0 @@
-ENVIRONMENT: "test"
-
-INPUT_SHAPE: [32, 32, 3]
-
-CATEGORIES_COUNT: 43
-CATEGORIES_IDS_TO_NAMES_CSV_PATH: "./data/signs_ids_to_names.csv"
-
-FLASK:
-
-  DEBUG: True
